Player now rotates and avoids jumpcolision
Jumpcolision while coliding with a wall will teleport player1 outside camera collideers.

When player flips, hurtcoliders and hitcoliders wont rotate, and some animations with displacement will displace player in negative axis. (This can be fixed with an if condition and + or - depoending on the animation displacement).

Player needs to be rotated over pivot, this is not happening now.